requirements.txt
celery==5.4.0
crispy-bootstrap5==2024.2
Django==5.0.4
django-environ==0.11.2
django-crispy-forms==2.1
djangorestframework==3.15.1
gunicorn==22.0.0
mysqlclient==2.2.4
pillow==10.3.0

